+++ ListlyApp/src/screens/CollectionsScreen.tsx
@@ -1,5 +1,6 @@
+import { LIST_VIEW_MODES } from '../constants/types';
 import ListsScreenBase from './ListsScreenBase';
 
 export default function CollectionsScreen() {
-  return <ListsScreenBase listsLayoutKey="collectionsLayout" collectionsLayoutKey="collectionsLayout" mode="collections" />;
+  return <ListsScreenBase listsLayoutKey="collectionsLayout" collectionsLayoutKey="collectionsLayout" mode={LIST_VIEW_MODES.collections} />;
 }