+++ ListlyApp/src/screens/HomeScreen.tsx
@@ -1,5 +1,6 @@
+import { LIST_VIEW_MODES } from '../constants/types';
 import ListsScreenBase from './ListsScreenBase';
 
 export default function HomeScreen() {
-  return <ListsScreenBase listsLayoutKey="homeListsLayout" collectionsLayoutKey="homeCollectionsLayout" mode="home" />;
+  return <ListsScreenBase listsLayoutKey="homeListsLayout" collectionsLayoutKey="homeCollectionsLayout" mode={LIST_VIEW_MODES.home} />;
 }